magfi-ingestor/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
feedparser==6.0.10
httpx==0.25.1
requests==2.31.0
openai==1.3.7
google-generativeai==0.3.0
apscheduler==3.10.4
